MainVCにControllerを導入し、Presenterから処理を分割する
