+++ iOSEngineerCodeCheck/MainPresenter.swift
@@ -3,7 +3,6 @@ import Foundation
 
 @MainActor
 final class MainPresenter {
-    private unowned let router: NavigationRouter
     private unowned let searcher: GitHubSearcher
 
     private let contentsSubject: CurrentValueSubject<[MainCellContent], Never> = .init([])
@@ -14,28 +13,11 @@ final class MainPresenter {
 
     private var cancellables: Set<AnyCancellable> = []
 
-    init(router: NavigationRouter, searcher: GitHubSearcher) {
-        self.router = router
+    init(searcher: GitHubSearcher) {
         self.searcher = searcher
 
         searcher.statePublisher.map(\.mainCellContents).subscribe(contentsSubject).store(
             in: &cancellables)
-    }
-
-    func search(word: String) {
-        searcher.search(word: word)
-    }
-
-    func cancelSearch() {
-        searcher.cancel()
-    }
-
-    func didAppear() {
-        router.mainDidAppear()
-    }
-
-    func showDetail(at index: Int) {
-        router.showDetail(.init(repositoryIndex: index))
     }
 }
